requirements.txt
streamlit
st-annotated-text